+++ CHANGES.rst
@@ -14,7 +14,7 @@ Changes:
 --------
 - Add `CLI` and `WeaverClient` support of ``logs``, ``exceptions`` and ``statistics`` retrieval.
 - Add `CLI` and `WeaverClient` support of `Job` search filtered by ``tags``, ``process`` and ``providers`` queries.
-- Add `CLI` and `WeaverClient` support of `Job` search filtered by multiple ``status`` values.
+- Add `CLI`, `WeaverClient` and `API` support of `Job` search filtered by multiple ``status`` values.
 - Adjust OpenAPI schema definitions for `Process` deployment to allow ``owsContext`` by itself without duplicated
   information that was required by mandatory ``executionUnit`` definition.